Game of life completed
game of life now completed in python.
